Add files via upload
Wins are detected in the top left sector only. automates between O and X with random positions currently. 
Next to add click to place markers. In the future dropped by a moving character (Bomberman style) or after a set number of squares moved (say 9) Also allow for rows and cols to be shifted (limitations on this to be decided - perhaps instead of placing a marker).